+++ crates/sncast/src/starknet_commands/get/mod.rs
@@ -1,6 +1,7 @@
 use clap::{Args, Subcommand};
 use sncast::helpers::configuration::CastConfig;
 use sncast::response::ui::UI;
+use std::process::ExitCode;
 
 pub mod balance;
 pub mod class_hash_at;
@@ -35,20 +36,16 @@ pub enum GetCommands {
     ClassHashAt(class_hash_at::ClassHashAt),
 }
 
-pub async fn get(get: Get, config: CastConfig, ui: &UI) -> anyhow::Result<()> {
+pub async fn get(get: Get, config: CastConfig, ui: &UI) -> anyhow::Result<ExitCode> {
     match get.command {
-        GetCommands::TxStatus(status) => tx_status::tx_status(status, config, ui).await?,
+        GetCommands::TxStatus(status) => tx_status::tx_status(status, config, ui).await,
 
-        GetCommands::Tx(tx) => transaction::transaction(tx, config, ui).await?,
+        GetCommands::Tx(tx) => transaction::transaction(tx, config, ui).await,
 
-        GetCommands::Balance(balance) => balance::balance(balance, config, ui).await?,
+        GetCommands::Balance(balance) => balance::balance(balance, config, ui).await,
 
-        GetCommands::Nonce(nonce) => nonce::nonce(nonce, config, ui).await?,
+        GetCommands::Nonce(nonce) => nonce::nonce(nonce, config, ui).await,
 
-        GetCommands::ClassHashAt(args) => {
-            class_hash_at::class_hash_at(args, config, ui).await?;
-        }
+        GetCommands::ClassHashAt(args) => class_hash_at::class_hash_at(args, config, ui).await,
     }
-
-    Ok(())
 }